Ensure that trust checks are run even when dpkg_conffile_prompt is False
The trust check is currently performed when the packages are downloaded
for inspection for a conffile prompt. If the user has setup apt so
that --force-confold/--force-confnew is passed this check will not
performed and there is a risk that unauthenticated packages get downloaded.

This branch moves the test if the package is trusted to an earlier part
of the code to ensure it will not be bypassed even if
dpkg_conffile_prompt() returns False.