display translations in proper locale
without this, translations would get encoded using the ascii codec
which would break with an infamous UnicodeEncodeError exception. by
using lgettext(), the proper encoding is automatically chosen on
output which seems to be a good practice here.

notice that i had to change from the string formatting operator to
arguments to the logging function in a specific case, otherwise i
would get the following exception:

TypeError: unsupported operand type(s) for %: 'bytes' and 'list'

This is a weird Python 3 specific exception, but it makes sense: the
string is encoded to bytes already and the formatting operator doesn't
want to take any chances. the proper way to fix this is to use the
builtin formatting of the logging function. i have done this in the
one place where i still had, but it's possible that i missed other possible places.