Speed up sanity check by filtering cache.
The apt.cache.Cache.get_changes() method returns only packages that are not marked as keep. The sanity check does not care for packages that are marked as keep, so it seems to make sense to filter the cache first to avoid unneeded loops.

On a test system with 140 upgradeable packages, this reduced the runtime of unattended-upgrade from 4m32s to 1m5s (without installation of packages).

This merge request was already done on the debian/jessie branch but I
would like to get this in the master branch as well.