Protect periodically run unattended-upgrades from being killed early
Add --with-install-on-shutdown option to unattended-upgrades-shutdown
and don't start u-u-s with it when it just have to stop and wait for
u-u.

The u-u.service is changed to wait for periodically ran u-u then
install all updates on shutdown.

The newly provided drop-in for apt-daily-upgrade.service prevents
the service from just being killed the added ExecStop=
uses u-u-s to stop the running u-u instead gracefully and showing
progress. The ExecStop= call does not pass --with-install-on-shutdown
thus no extra package installation is started by u-u-s.

Depend on apt >= 1.4.2 which split apt-daily-upgrade.service from
apt-daily.service.

LP: #1690980