Run u-u-s only when running on AC power
When InstallOnShutdown it set to true u-u-s is running the upgrades step
which is risky when a laptop's battery is almost depleted because the
remaining power may not be enough to install all upgrades leaving the
system in an inconsistent state.
When InstallOnShutdown is false, then apt-daily-upgrade.service performs
the upgrade in recent apt versions and it does not run on battery power
thus u-u-s does not have to stop it. Apt-daily-upgrade.service also runs
u-u in a way that ensures graceful termination even without u-u-s.