unattended-upgrades: fix Unlocked context manager
The Unlocked context manager did correctly unlock but did not
reacquire the lock which means that in minimal-upgrade step
mode it is possible to run apt code without a lock. If something
else (like landscape, apt, synaptic, packagekit) locks the cache
in the meantime this will work and u-u will get dpkg errors
because dpkg will not be able to perform its operations. It is
less of an issue in non-minimal mode, but even then the auto-remove
step may fail in this way.

We should also SRU this fix.