Stop using ActionGroups, they interfere with apt.Cache.clear()
causing all autoremovable packages to be handled as newly autoremovable ones
and be removed by default. Dropping ActionGroup usage does not slow down the
most frequent case of not having anything to upgrade and when ther are
packages to upgrade the gain is small compared to the actual package
installation.

Also collect autoremovable packages before adjusting candidates because that
also changed .is_auto_removable attribute of some of them.

LP: #1803749
Closes: #910874