test_dev_release: Fix and enable test
Enable the test, this requires overriding the LOCK_FILE to
not use the global one. Also set some more config options
needed for travis, and abstract the config writing into a
method so we don't repeat ourselves if we add more test
cases.

Since we are sharing root.untrusted with another test,
we also have to make sure not to replace it's apt.conf,
so we just move it out of the way in setUp and back in
tearDown.