Adjust only direct dependencies when falling back to adjusting more packages
Adjusting all transitive dependencies is found to slow down upgrades too much
when many packages were upgradable but not with adjustments taking place.
If a package upgrade to be installed by unattended-upgrades requires adjusting
a transitive but not direct dependency, please add this dependency as a direct
dependency of the package to unlock the upgrade.

Closes: #935653
Fixes: #229
LP: #1848357