unattended-upgrade: add CM for toggling process priority
Add a context manager for wrapping code that would benefit
with a different process execution priority, only during that
block of code execution.

It is done in response of a suggestion in the code itself to wrap
some logic in such a context manager.

Then that logic is wrapped in it.