debian/tests: pass `-Zgzip` to dpkg-deb to avoid errors on stable
When building on sid dpkg-deb uses zstd now by default. However this
is not yet supported on ubuntu 22.04 so pass `-Zgzip` to dpkg-deb
to create a compatible deb so that the  `upgrade-all-secrity`
test can work.

Sadly this needs to be (slightly) hacked because there is no way
right now to override this via an env var in dpkg AFACIT, so a new
UU_BUILDDEB_ARGS is added.